requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# NLP and ML
spacy==3.7.2
transformers==4.36.2
sentence-transformers==2.3.1
torch==2.1.2
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
langdetect==1.0.9
rank-bm25==0.2.2

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25
alembic==1.13.1

# LLM Integration
llama-cpp-python==0.2.27  # For local Llama with Metal support
openai==1.10.0  # For API fallback
anthropic==0.8.1  # Alternative API option

# Utilities
loguru==0.7.2
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.26.0
tenacity==8.2.3

# Monitoring and Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For testing FastAPI

# Development
black==23.12.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0
